chore: slice workflow refresh draft hook
