+++ web/app/components/workflow-app/components/workflow-main.tsx
@@ -8,6 +8,7 @@ import type { WorkflowProps } from '@/app/components/workflow'
 import WorkflowChildren from './workflow-children'
 import {
   useNodesSyncDraft,
+  useWorkflowRefreshDraft,
   useWorkflowRun,
   useWorkflowStartRun,
 } from '../hooks'
@@ -32,6 +33,7 @@ const WorkflowMain = ({
     doSyncWorkflowDraft,
     syncWorkflowDraftWhenPageClose,
   } = useNodesSyncDraft()
+  const { handleRefreshWorkflowDraft } = useWorkflowRefreshDraft()
   const {
     handleBackupDraft,
     handleLoadBackupDraft,
@@ -49,6 +51,7 @@ const WorkflowMain = ({
     return {
       syncWorkflowDraftWhenPageClose,
       doSyncWorkflowDraft,
+      handleRefreshWorkflowDraft,
       handleBackupDraft,
       handleLoadBackupDraft,
       handleRestoreFromPublishedWorkflow,
@@ -61,6 +64,7 @@ const WorkflowMain = ({
   }, [
     syncWorkflowDraftWhenPageClose,
     doSyncWorkflowDraft,
+    handleRefreshWorkflowDraft,
     handleBackupDraft,
     handleLoadBackupDraft,
     handleRestoreFromPublishedWorkflow,
