+++ web/app/components/workflow-app/hooks/index.ts
@@ -4,3 +4,4 @@ export * from './use-nodes-sync-draft'
 export * from './use-workflow-run'
 export * from './use-workflow-start-run'
 export * from './use-is-chat-mode'
+export * from './use-workflow-refresh-draft'
