+++ web/app/components/workflow/hooks/index.ts
@@ -16,3 +16,4 @@ export * from './use-shortcuts'
 export * from './use-workflow-interactions'
 export * from './use-workflow-mode'
 export * from './use-format-time-from-now'
+export * from './use-workflow-refresh-draft'
