+++ web/app/components/workflow/hooks/use-workflow-refresh-draft.ts
@@ -1,0 +1,9 @@
+import { useHooksStore } from '@/app/components/workflow/hooks-store'
+
+export const useWorkflowRefreshDraft = () => {
+  const handleRefreshWorkflowDraft = useHooksStore(s => s.handleRefreshWorkflowDraft)
+
+  return {
+    handleRefreshWorkflowDraft,
+  }
+}
